Move `create_remove` to `commands`
