--- packages/engine/src/datastore/table/mod.rs
@@ -1,4 +1,0 @@
-//! TODO, rework when we do separation of crates, probably remove table module
-//! This is about a single simulation's data, conceptually we have Tables for agent state, context,
-//! message state
-pub mod create_remove;
