+++ packages/engine/src/simulation/error.rs
@@ -1,4 +1,3 @@
-use stateful::agent::Agent;
 use thiserror::Error as ThisError;
 use tokio::sync::mpsc::error::SendError;
 
@@ -15,38 +14,14 @@ pub enum Error {
     #[error("Execution error: {0}")]
     Execution(#[from] execution::Error),
 
-    #[error("Datastore Error: {0}")]
-    DataStore(#[from] crate::datastore::Error),
+    #[error("Command Error: {0}")]
+    Command(#[from] crate::simulation::command::Error),
 
     #[error("Controller Error: {0}")]
     Controller(#[from] super::controller::Error),
 
     #[error("Tokio oneshot recv: {0}")]
     TokioOneshotRecv(#[from] tokio::sync::oneshot::error::RecvError),
-
-    #[error("Unexpected message to hash with type {message_type}")]
-    UnexpectedSystemMessage { message_type: String },
-
-    #[error("Serde Error: {0}")]
-    Serde(#[from] serde_json::Error),
-
-    #[error(
-        "Received an incorrect `remove_agent` message: {0}. Valid examples: 1) {{\"agent_id\": \
-         \"b2387514-e76a-4695-9831-8d9ac6254468\"}}, 2) None/null 3) {{}}, 4) \"\""
-    )]
-    RemoveAgentMessage(String),
-
-    #[error(
-        "Error parsing `create_agent` message payload, expected valid agent state, got error: \
-         {0:?}. Payload was: {1:?}"
-    )]
-    CreateAgentPayload(serde_json::error::Error, String),
-
-    #[error(
-        "`create_agent` message has field \"{0}\" without respective field existing\nDetails: \
-         {1:?}"
-    )]
-    CreateAgentField(String, Agent),
 
     #[error("Tokio Join Error: {0}")]
     TokioJoin(#[from] tokio::task::JoinError),
